+++ CHANGELOG.md
@@ -15,3 +15,9 @@
 ## 0.0.4
 
 - remove msg
+
+## 0.0.5
+
+- cleanup
+- remove lines with spaces only too
+- better handling for start of doc empty lines